+++ src/test/java/roomescape/controller/AdminControllerTest.java
@@ -1,0 +1,31 @@
+package roomescape.controller;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+class AdminControllerTest {
+
+    @DisplayName("어드민 메인 페이지 요청 테스트")
+    @Test
+    void adminMainPageTest() {
+        RestAssured.given().log().all()
+                .when().get("/admin")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("어드민 예약 페이지 요청 테스트")
+    @Test
+    void adminReservationPageTest() {
+        RestAssured.given().log().all()
+                .when().get("/admin/reservation")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+}
