+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 import static org.hamcrest.Matchers.is;
 
@@ -13,25 +13,7 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class MissionStepTest {
-
-    @DisplayName("어드민 메인 페이지 요청 테스트")
-    @Test
-    void adminMainPageTest() {
-        RestAssured.given().log().all()
-                .when().get("/admin")
-                .then().log().all()
-                .statusCode(200);
-    }
-
-    @DisplayName("어드민 예약 페이지 요청 테스트")
-    @Test
-    void adminReservationPageTest() {
-        RestAssured.given().log().all()
-                .when().get("/admin/reservation")
-                .then().log().all()
-                .statusCode(200);
-    }
+class ReservationControllerTest {
 
     @DisplayName("전체 예약 정보 요청 테스트")
     @Test
